Added Migration and updated database

+++ VatClinic.Data/Entities/Resolve.cs
@@ -1,13 +1,17 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Text;
 
 namespace VatClinic.Data.Entities
 {
     public class Resolve
     {
-        public int AilmentId { get; set; }
-        public int MedicationId { get; set; }
+       
+        public long AilmentId { get; set; }
+
+        
+        public long MedicationId { get; set; }
 
         public Ailment Ailment { get; set; }
         public Medication Medication { get; set; }
